backend/api_v1/serializers.py
from rest_framework import serializers
from products.models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        fields = (
            'id',
            'title',
            'main_picture',
            'price',
        )
        model = Product
